ChatbotService.API/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.4
google-generativeai==0.8.3
